Add conditional logic for navbar rendering
