+++ agile_aprons/http/user.http
@@ -1,12 +1,7 @@
 @url = http://localhost:8080/api/user
 
-GET {{url}} HTTP/1.1
+GET {{url}}/1 HTTP/1.1
 Content-Type: application/json
-# Authorization: Bearer 
-
-{
-  "id": 4
-}
 
 ###
 PUT {{url}} HTTP/1.1
