+++ clients/client-web/src/components/views/Profile.jsx
@@ -2,11 +2,13 @@ import { useState, useContext, useEffect } from 'react';
 import ProfileCard from '../ProfileCard';
 import AuthContext from '../../contexts/AuthContext';
 import { findUser } from '../../service/userApi';
+import { useNavigate } from 'react-router-dom';
 
 const Profile = () => {
     const [userProps, setUserProps] = useState({});
     
     const auth = useContext(AuthContext);
+    const navigate = useNavigate();
     useEffect(() => {
         findUser()
             .then(data => {
@@ -14,13 +16,19 @@ const Profile = () => {
                 auth.userProps = {...data};
             })
             .catch(err => console.log(err));
-        console.log(userProps)
     }, []);
+    
+    useEffect(() => {
+        if (!auth.isLoggedIn()) {
+            navigate('/');
+        }
+    }, [auth]);
     
     return (
         <>
             <ProfileCard 
-                // name={}
+                // imgUrl={}
+                name={userProps.displayName}
             />
         </>
     )
